+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+
+- `shutdown-after` cli parameter and config option now properly shuts down
+  server after N seconds with no connections
+
 ## [0.17.5] - 2022-08-18
 ### Fixed